// src/app/taf-fiscal/models/item-table-resources-received-by-the-sports-association-validation.ts
export interface ItemTableResourcesReceivedByTheSportsAssociationValidation {
  branchId: string;
  branch: string;
  branchTaxNumber?: string;
  taxNumber?: string;
  taxNumberFormated: string;
  company: string;
  totalInvoice: number;
  totalTransferAmount: number;
  totalGrossValue: number;
  totalTaxBase: number;
  totalTaxes: number;
  key: string;
  status: string;
  errors: string;
  keyValidationErrors?: string;
}