// src/app/taf-fiscal/models/item-table-social-security-contribution-validation.ts
export interface ItemTableSocialSecurityContributionValidation {
  status?: string;
  typeOfInscription?: string;
  companyTaxNumber?: string;
  taxNumberFormated?: string;
  totalInvoice?: number;
  totalGrossValue?: number;
  sociaSecurityContributionValue?: number;
  sociaSecurityContributionValueSuspended?: number;
  key?: string;
  branchId?: string;
  existProcId?: string;
}